bugs fixed 2

+++ app/api/client/[id]/route.ts
@@ -4,25 +4,36 @@ import { NextResponse } from "next/server";
 import { eq } from "drizzle-orm";
 import { clients, invoices } from "@/db/schema";
 import { UUID } from "crypto";
+import { error } from "console";
+
+type RouteParams = Promise<{ id:string }>
 
 export async function DELETE(
   req: Request,
-  { params }: { params: { id: UUID } },
+  { params }: { params: RouteParams },
 ) {
   const { id } = await params;
   try {
-    await db.delete(clients).where(eq(clients.id, id));
-    return NextResponse.json({ success: true }, { status: 200 });
+    if(!id){
+
+      throw new Error('id not found')
+
+    }
+      await db.delete(clients).where(eq(clients.id, id));
+      return NextResponse.json({ success: true }, { status: 200 });
   } catch (error) {
     console.error("Error deleting client:", error);
     return NextResponse.json({ success: false, error: error }, { status: 500 });
   }
 }
 
-export async function PUT(req: Request, { params }: { params: { id: UUID } }) {
+export async function PUT(req: Request, { params }: { params: RouteParams } ) {
   const { id } = await params;
   try {
     const body = await req.json();
+    if(!id){
+      throw new Error('id not found')
+    }
     const updated = await db
       .update(clients)
       .set({
